+++ src/InteractionRecorder.tsx
@@ -8,7 +8,7 @@ import { SaveStoryButton } from './SaveStory';
 import { combineInteractions } from './codegen/combine-interactions';
 import { convertInteractionsToCode } from './codegen/interactions-to-code';
 import { EVENTS } from './constants';
-import { type Interaction, useIsRecording, useRecorderState } from './state';
+import { type Interaction, useInteractions, useIsRecording } from './state';
 import {
 	CodeBlocksWrapper,
 	Container,
@@ -22,25 +22,22 @@ import {
 } from './styles';
 
 export const InteractionRecorder = () => {
-	const [{ interactions }, setState] = useRecorderState();
+	const [interactions, setInteractions] = useInteractions();
 	const [isRecording, setIsRecording] = useIsRecording();
 
 	useChannel({
 		[EVENTS.INTERACTION]: (interaction: Interaction) => {
-			setState((state) => ({
-				...state,
-				interactions: combineInteractions(interaction, state.interactions),
-			}));
+			setInteractions((prevInteractions) =>
+				JSON.stringify(
+					combineInteractions(interaction, JSON.parse(prevInteractions)),
+				),
+			);
 		},
 	});
 
 	const toggleRecording = () => setIsRecording(!isRecording);
 
-	const resetEvents = () =>
-		setState((state) => ({
-			...state,
-			interactions: [],
-		}));
+	const resetInteractions = () => setInteractions(() => JSON.stringify([]));
 
 	const api = useStorybookApi();
 
@@ -48,7 +45,7 @@ export const InteractionRecorder = () => {
 
 	// biome-ignore lint/correctness/useExhaustiveDependencies: reset events & recording when story changes
 	useEffect(() => {
-		resetEvents();
+		resetInteractions();
 	}, [storyData?.id]);
 
 	const hasTypescript = ['.ts', '.tsx'].some((ext) =>
@@ -57,7 +54,8 @@ export const InteractionRecorder = () => {
 
 	const [debouncedInteractions] = useDebounce(interactions, 100);
 	const code = useMemo(
-		() => convertInteractionsToCode(debouncedInteractions, hasTypescript),
+		() =>
+			convertInteractionsToCode(JSON.parse(debouncedInteractions), hasTypescript),
 		[debouncedInteractions, hasTypescript],
 	);
 
@@ -102,7 +100,7 @@ export const InteractionRecorder = () => {
 								{isRecording ? 'Stop' : 'Start'} recording
 							</StyledButton>
 
-							<StyledButton onClick={resetEvents} disabled={!code.play.length}>
+							<StyledButton onClick={resetInteractions} disabled={!code.play.length}>
 								<DeleteIcon />
 								Reset
 							</StyledButton>